[CL-677] Create separate component for the iframe
